Formalize deprecation of token_api.list_tokens
The token_api.list_tokens method should really only be used internally
to keystone.  This has been the case (as per the docstring) since the
Havana development cycle.  This patchset formalizes the deprecation and
slates the public access to token_api.list_tokens to be removed once
Icehouse ships.

bp: deprecated-as-of-icehouse
Change-Id: I30a066d3775e67fd4a84f07dc351fcfd62751045