Remove unused token.valid index
Because valid is only ever used in conjunction with expires, we do not
need it to have its own index now that there is a covering compound
index for expires and valid.

Note that he expires index is still useful alone for purging old tokens
as we do not filter for valid in that case.

Change-Id: Ib8ac2ab740b480f8ab29c07e66247f4fce15150b