Add assertSetEqual to base test class
Implement and add assertSetEqual to the base test class so that the
functionaly of set-specific comparison can be used in python 2.6.
Once python 2.6 support is deprecated completly, this new method
can be deleted and the tests will fall back onto the upstream version
of this test.  This is a copy of the method implemented on the TestCase
super class in python 2.7 with minor changes to conform to OpenStack
Hacking.

bp: dogpile-kvs-backends
Change-Id: I76a7d9a10b30ed6b0ac0aef4e5dd522cafb321e2