Reduces memory utilization during test runs
It turns out that when the Python test framework runs it will create an
instance of TestCase (actually the subclass containing our tests) of
each test in the TestCase. These instances don't get cleaned up until
the very end of the test run. This means that if there are lots of tests
and/or lots of instance variables created in the setUp() you use lots of
memory.

Change-Id: Ie9451feafd47d4a4b9a0ff6d8459b45a1645d1ee