Do not update password when updating grants in Assignment KVS
Assignment KVS backend no longer performs an update that causes
the user's password to be updated when updating grants. This
potentially would have caused any/all grant updates to set the
user's password to the current password hash, effectively making
it impossible to login with the user. The longer-term fix
should ensure that the backend does not need to update the user
itself from the driver level.

Change-Id: I74ffc643020fa077c39a0cf16847e397a0c24139
Closes-bug: #1265887