Allow event callback registration for arbitrary resource types
In some cases events should still be able to be registered for
resource types that have not been explicitly defined due to an
extension not being loaded or active. The if-check to raise
a ValueError if the ResourceType has not been explicitly
defined by a @dependancy.provides line has been removed.

Change-Id: I9fb368c5734a4a560a4bdd97d4467e2f76ca0e5c