Tests initialize database
Many tests (for example the backend_kvs tests) failed to run
individually. This is because the tests made the assumption that
because the system is configured for the KVS (or non-SQL) backend
that the database isn't required. The problem with this assumption
is that different backends can use different storage mechanisms,
and the credential backend only supports SQL. So any test that
requires that backends are loaded is required to initialize the
database.

Change-Id: I109b83bf512598952346bef01dc33c57802c7f60