Removes use of fake_notify and fixes notify test
The test against keystone.notifications._send_notification was badly
broken, and the stubbed fake_notify method was masking this issue.

The operation and resource parameters to _send_notification() were
mistakenly reversed in the test case's call to that method, but because
the fake_notify() method was not looking for the parameters in the right
way, the test was failing.

This patch removes the unnecessary use of mox stubout and replaces with
a simple call to mock.patch() and a follow-up assertion that the mocked
notify() call was being passed expected arguments.

Change-Id: I1f3e66c7b92e176520c7627151a02d4b35594aea
Blueprint: oslo-messaging