Cleanup oauth tests
In a lot of places this code was using dict.get() where [] makes
more sense because the code then assumes that the field was
present and had a value. Using .get() just means that if the code
changes the error would be an unexpected TypeError on a different
line than where the problem actually was.

Also, building of simple url strings was simplified to remove the
dict so that they could be kept on a single line.

Change-Id: I98b5631e02fe1456f0f583c4ec55c9ac0f8c4c35