Remove duplicated cms file
Co-Authored-By: David Stanek (dstanek@dstanek.com)

Until recently, the Unit tests for the Keystone client meant that
we could not use the most recent version of the client code inside
the server.  That lead to code duplication between client and server
for new features.  Now, we are only running unit tests against the
master branch of the client.  More and more common code can be moved
to the client without duplication.

cms was duplicated between the server and client.
Use the version from the client.

Tests for error handling problems in cms

Change-Id: Ieed0ba29f55216c5a6819bab7d9b862f2aebbeb0