Always hash passwords on their way into the DB
Way back in ed793ad5 a feature was added to detect if a password was
already hashed. If it looked like it was already hashed the password
was put into the database as is. It appears that this was added to
support a migration that was using the identity API to store users
instead of direct database inserts.

Change-Id: I37b575e4db2fa7090c5171b8bc32cf2c57ad6e03
Fixes-bug: #1279849