Initial implementation of validator
This is the initial implementation of a jsonschema type
validator for the Keystone API. Something very similar to this is used
in Nova V3 for validation. By using jsonschema to validate API requests,
we can make parameters that are passed in fit the criteria we allow and
what is specified in the Identify API spec. This will also allow us to
validate parameters by wrapping the method that needs validation.

bp: api-validation
Change-Id: I1e1dc8e5ac3ad766f05444b16d56a22c89602b9f