Add tests for endpoint enabled.
There were no tests that show what happens when you create or
update the 'enabled' field in endpoints with different types of
values. There were also no tests that show what happens when the
'enabled' field isn't specified when an endpoint is created.

The new tests show that you can set the 'enabled' field to
any type of value even though the identity API specification says
that it's a Boolean. The new tests also show that if the 'enabled'
field isn't set explicitly then it's not returned on the response.

Change-Id: Iadc2a0ed3d6aaaa89b82f4caa8829ecd1d8f2712
Related-Bug: #1282266