Fixup region description uniqueness
Add a db migration for relaxing the uniqueness constraint on the
'description' column in the region table.

The migration will only occur if there is a unique index constraint
on the 'description' table. The previous migration of 037 has
had the unique index dropped on table creation. 043 will only be
run in the case of a deployment tracking closer to master than
the official releases.

Related-Bug: 1284972
Change-Id: I8198aec2fea4ff6a88938696499a60ebc0b16a47